+++ djangocms_navigation/admin.py
@@ -305,7 +305,10 @@ class MenuItemAdmin(TreeAdmin):
 
     class Media:
         css = {
-            "all": ("djangocms_versioning/css/actions.css",)
+            "all": (
+                "djangocms_versioning/css/actions.css",
+                "djangocms_navigation/css/navigation_admin_changelist.css",
+            )
         }
 
     def get_urls(self):