+++ djangocms_navigation/admin.py
@@ -401,8 +401,22 @@ class MenuItemAdmin(TreeAdmin):
         Collect rendered actions from implemented methods and return as list
         """
         return [
+            self._get_edit_link,
             self._get_delete_link,
         ]
+
+    def _get_edit_link(self, obj, request, disabled=False):
+        app, model = self.model._meta.app_label, self.model._meta.model_name
+
+        edit_url = reverse(
+            "admin:{app}_{model}_change".format(app=app, model=model),
+            args=[request.menu_content_id, obj.id]
+        )
+
+        return render_to_string(
+            "djangocms_versioning/admin/edit_icon.html",
+            {"edit_url": edit_url, "disabled": disabled, "object_id": obj.id}
+        )
 
     def _get_delete_link(self, obj, request, disabled=False):
         app, model = self.model._meta.app_label, self.model._meta.model_name