+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Unreleased
 
 1.0.0 (2022-02-18)
 ===================
+* MenuItem edit button added as list action button
 * MenuItem delete moved from dropdown action to individual delete action button
 * Python 3.8, 3.9 support added
 * Django 3.0, 3.1 and 3.2 support added