+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 
 Unreleased
 ==========
+* feat: MenuItem edit button added as list action button
 * feat: Hide the MenuItem boolean false icon for `Hide in Menu` and `Soft Root` fields.
 * feat: MenuItem changelist admin, remove the option to sort by `Hide In Menu` and `Soft Root`.